+++ StripeCryptoOnramp/StripeCryptoOnramp/Source/StripeCryptoOnramp.swift
@@ -1,0 +1,10 @@
+//
+//  StripeCryptoOnramp.swift
+//  StripeCryptoOnramp
+//
+//  Created by Michael Liberatore on 7/9/25.
+//
+
+import Foundation
+
+// Placeholder file for framework structure setup PR.