+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetAppearance.swift
@@ -315,7 +315,7 @@ public extension PaymentSheet.Appearance {
                 case flatWithCheckmark
                 /// A flat style with a chevron
                 /// Note that EmbeddedPaymentElement.Configuration.RowSelectionBehavior must be set to `immediateAction` to use this style.
-                case flatWithChevron
+                case flatWithDisclosure
             }
 
             /// The display style of the row
@@ -324,6 +324,14 @@ public extension PaymentSheet.Appearance {
             /// Additional vertical insets applied to a payment method row
             /// - Note: Increasing this value increases the height of each row
             public var additionalInsets: CGFloat = 6.0
+
+            /// The font of the title in a payment method row e.g. "New card"
+            /// - Note: If `nil`, uses a default font based on `appearance.font`
+            public var titleFont: UIFont?
+
+            /// Controls the padding around the payment method icon.
+            /// - Note: The top and bottom margins are ignored; use `additionalInsets` to control the height of the row.
+            public var paymentMethodIconLayoutMargins: NSDirectionalEdgeInsets = .init(top: 0, leading: 12, bottom: 0, trailing: 12)
 
             /// Appearance settings for the flat style
             public var flat: Flat = Flat()
@@ -341,7 +349,7 @@ public extension PaymentSheet.Appearance {
                 public var separatorColor: UIColor?
 
                 /// The insets of the separator line between rows
-                /// - Note: If `nil`, defaults to `UIEdgeInsets(top: 0, left: 30, bottom: 0, right: 0)` for style of `flatWithRadio` and to `UIEdgeInsets.zero` for styles of `flatWithCheckmark` and `flatWithChevron`.
+                /// - Note: If `nil`, defaults to `UIEdgeInsets(top: 0, left: 30, bottom: 0, right: 0)` for style of `flatWithRadio` and to `UIEdgeInsets.zero` for styles of `flatWithCheckmark` and `flatWithDisclosure`.
                 public var separatorInsets: UIEdgeInsets?
 
                 /// Determines if the top separator is visible at the top of the Embedded Mobile Payment Element
@@ -356,8 +364,8 @@ public extension PaymentSheet.Appearance {
                 /// Appearance settings for the checkmark
                 public var checkmark: Checkmark = Checkmark()
 
-                /// Appearance settings for the chevron
-                public var chevron: Chevron = Chevron()
+                /// Appearance settings for the disclosure (by default, a chevron)
+                public var disclosure: Disclosure = Disclosure()
 
                 /// Describes the appearance of the radio button
                 public struct Radio: Equatable {
@@ -377,11 +385,16 @@ public extension PaymentSheet.Appearance {
                     public var color: UIColor?
                 }
 
-                /// Describes the appearance of the chevron
+                /// Describes the appearance of the disclosure (by default, a chevron)
                 /// Note that EmbeddedPaymentElement.Configuration.RowSelectionBehavior must be set to `immediateAction` to use this style.
-                public struct Chevron: Equatable {
-                    /// The color of the chevron icon
+                public struct Disclosure: Equatable {
+                    /// The color of the default chevron icon
                     public var color: UIColor = .systemGray
+
+                    /// The image displayed on the right of the row.
+                    /// - Note: If `nil` (the default), a chevron is displayed.
+                    @_spi(CustomEmbeddedDisclosureImagePreview)
+                    public var disclosureImage: UIImage?
                 }
             }