+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethod.swift
@@ -121,6 +121,8 @@ public class STPPaymentMethod: NSObject, STPAPIResponseDecodable {
 
     /// The payment details of a PaymentMethod that was created using Link.
     @_spi(STP) public var linkPaymentDetails: LinkPaymentDetails?
+    /// Whether this payment method is coming from Link.
+    @_spi(STP) public var isLinkPassthroughMode: Bool = false
 
     @_spi(STP) public var isLinkPaymentMethod: Bool {
         linkPaymentDetails != nil